Move structure from `testing` dir to main dir
