+++ votepy/algorithms/base_algorithm.py
