+++ votepy/algorithms/generic_brute_force.py
@@ -1,6 +1,6 @@
 from votepy.ordinal_election import OrdinalElection
-from votepy.testing.algorithms.base_algorithm import BaseAlgorithm
-from votepy.testing.structure.structure import algo
+from votepy.algorithms.base_algorithm import BaseAlgorithm
+from votepy.structure.structure import algo
 
 from itertools import combinations
 from typing import Callable, Iterable
