+++ votepy/solve.py
@@ -1,8 +1,8 @@
 from re import I
 from typing import Callable, Iterable, Union
 from votepy.ordinal_election import OrdinalElection
-from votepy.testing.algorithms.base_algorithm import BaseAlgorithm
-from votepy.testing.structure.structure import get_algorithm, get_default_algorithm, get_implementation
+from votepy.algorithms.base_algorithm import BaseAlgorithm
+from votepy.structure.structure import get_algorithm, get_default_algorithm, get_implementation
 
 
 def solve(rule: Union[Callable, str], voting: Union[OrdinalElection, list[int]], size_of_committee: int, *rule_args, algorithm: BaseAlgorithm = None, **rule_kwargs) -> list[int]:
@@ -21,7 +21,7 @@ def solve(rule: Union[Callable, str], voting: Union[OrdinalElection, list[int]],
         -> list[int]: List of chosen candidates
 
     ## Examples
-    >>> from votepy.testing.rules.k_borda import k_borda
+    >>> from votepy.rules.k_borda import k_borda
     >>> solve(k_borda, [
     ...     [0, 1, 2, 3],
     ...     [3, 2, 1, 0],
